rework slots to remove comboboxbase component
