+++ packages/react-components/react-combobox/src/components/Combobox/Combobox.types.ts
@@ -1,36 +1,27 @@
-import * as React from 'react';
 import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
-import type { ComboboxContextValue } from '../../contexts/ComboboxContext';
-import type { ComboboxBaseProps, ComboboxBaseSlots, ComboboxBaseState } from '../ComboboxBase/ComboboxBase.types';
-import { ComboboxInput } from '../ComboboxInput/ComboboxInput';
+import type { ComboboxBaseProps, ComboboxBaseState } from '../../ComboboxBase/ComboboxBase.types';
+import { Listbox } from '../Listbox/Listbox';
 
-export type ComboboxSlots = Omit<ComboboxBaseSlots, 'input'> & {
-  input: NonNullable<Slot<typeof ComboboxInput>>;
+export type ComboboxSlots = {
+  /* The root combobox slot */
+  root: NonNullable<Slot<'div'>>;
+
+  /* The dropdown listbox slot */
+  listbox: NonNullable<Slot<typeof Listbox>>;
+
+  /* The primary slot, the element with role="combobox" */
+  input: NonNullable<Slot<'input'>>;
 };
 
 /**
  * Combobox Props
  */
-export type ComboboxProps = ComponentProps<ComboboxSlots> & Omit<ComboboxBaseProps, 'input'>;
+export type ComboboxProps = Omit<ComponentProps<Partial<ComboboxSlots>, 'input'>, 'children' | 'size'> &
+  ComboboxBaseProps & {
+    children: React.ReactNode;
+  };
 
 /**
  * State used in rendering Combobox
  */
-export type ComboboxState = Omit<ComboboxBaseState, 'components' | 'input'> & ComponentState<ComboboxSlots>;
-
-export type ComboboxContextValues = {
-  combobox: ComboboxContextValue;
-};
-
-/**
- * Data for the Combobox onOpenChange event.
- */
-export type ComboboxOpenChangeData = {
-  open: boolean;
-};
-
-/* Possible event types for onOpen */
-export type ComboboxOpenEvents =
-  | React.FocusEvent<HTMLElement>
-  | React.KeyboardEvent<HTMLElement>
-  | React.MouseEvent<HTMLElement>;
+export type ComboboxState = ComponentState<ComboboxSlots> & ComboboxBaseState;
