+++ packages/react-components/react-combobox/src/components/Combobox/renderCombobox.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
 import { Portal } from '@fluentui/react-portal';
 import { getSlots } from '@fluentui/react-utilities';
-import type { ComboboxContextValues, ComboboxState, ComboboxSlots } from './Combobox.types';
+import type { ComboboxContextValues } from '../../ComboboxBase/ComboboxBase.types';
 import { ComboboxContext } from '../../contexts/ComboboxContext';
+import type { ComboboxState, ComboboxSlots } from './Combobox.types';
 
 /**
  * Render the final JSX of Combobox
@@ -16,7 +17,7 @@ export const renderCombobox_unstable = (state: ComboboxState, contextValues: Com
   return (
     <slots.root {...slotProps.root}>
       <ComboboxContext.Provider value={contextValues.combobox}>
-        <slots.trigger {...slotProps.trigger} />
+        <slots.input {...slotProps.input} />
         {state.open ? popup : null}
       </ComboboxContext.Provider>
     </slots.root>
