+++ packages/react-components/react-combobox/src/components/Combobox/useComboboxStyles.ts
@@ -1,6 +1,5 @@
+import { makeStyles, mergeClasses } from '@griffel/react';
 import { SlotClassNames } from '@fluentui/react-utilities';
-import { ComboboxBaseState } from '../ComboboxBase/ComboboxBase.types';
-import { useComboboxBaseStyles_unstable } from '../ComboboxBase/useComboboxBaseStyles';
 import type { ComboboxSlots, ComboboxState } from './Combobox.types';
 
 export const comboboxClassNames: SlotClassNames<ComboboxSlots> = {
@@ -10,8 +9,22 @@ export const comboboxClassNames: SlotClassNames<ComboboxSlots> = {
 };
 
 /**
+ * Styles for the root slot
+ */
+const useStyles = makeStyles({
+  root: {},
+  listbox: {},
+  input: {},
+});
+
+/**
  * Apply styling to the Combobox slots based on the state
  */
 export const useComboboxStyles_unstable = (state: ComboboxState): ComboboxState => {
-  return useComboboxBaseStyles_unstable(state as ComboboxBaseState) as ComboboxState;
+  const styles = useStyles();
+  state.root.className = mergeClasses(comboboxClassNames.root, styles.root, state.root.className);
+  state.listbox.className = mergeClasses(comboboxClassNames.listbox, styles.listbox, state.listbox.className);
+  state.input.className = mergeClasses(comboboxClassNames.input, styles.input, state.input.className);
+
+  return state;
 };
