--- packages/react-components/react-combobox/src/components/ComboboxBase/index.ts
@@ -1,5 +1,0 @@
-export * from './ComboboxBase';
-export * from './ComboboxBase.types';
-export * from './renderComboboxBase';
-export * from './useComboboxBase';
-export * from './useComboboxBaseStyles';
