--- packages/react-components/react-combobox/src/components/ComboboxBase/useComboboxBase.ts
@@ -1,237 +1,0 @@
-import * as React from 'react';
-import { Offset, resolvePositioningShorthand, usePopper } from '@fluentui/react-positioning';
-import {
-  getPartitionedNativeProps,
-  resolveShorthand,
-  useControllableState,
-  useFirstMount,
-  useId,
-  useMergedRefs,
-} from '@fluentui/react-utilities';
-import { getDropdownActionFromKey, getIndexFromAction } from '../../utils/dropdownKeyActions';
-import { useOptionCollection } from '../../utils/useOptionCollection';
-import { OptionValue } from '../../utils/OptionCollection.types';
-import { useSelection } from '../../utils/useSelection';
-import { Listbox } from '../Listbox/Listbox';
-import type { ComboboxBaseProps, ComboboxBaseState, ComboboxOpenEvents } from './ComboboxBase.types';
-
-/**
- * Create the state required to render Combobox.
- *
- * The returned state can be modified with hooks such as useComboboxStyles_unstable,
- * before being passed to renderCombobox_unstable.
- *
- * @param props - props from this instance of Combobox
- * @param ref - reference to root HTMLElement of Combobox
- */
-export const useComboboxBase_unstable = (
-  props: ComboboxBaseProps,
-  ref: React.RefObject<HTMLButtonElement | HTMLInputElement>,
-): ComboboxBaseState => {
-  const optionCollection = useOptionCollection(props.children);
-
-  const {
-    appearance = 'outline',
-    inline = false,
-    multiselect,
-    onOpenChange,
-    placeholder,
-    positioning,
-    size = 'medium',
-  } = props;
-  const {
-    options,
-    collectionData: { count, getOptionAtIndex, getIndexOfKey, getOptionByKey },
-  } = optionCollection;
-
-  const [activeOption, setActiveOption] = React.useState<OptionValue | undefined>();
-  const { selectedOptions, selectOption } = useSelection(props);
-  const [open, setOpenState] = useControllableState({
-    state: props.open,
-    defaultState: props.defaultOpen,
-    initialState: false,
-  });
-
-  // handle trigger focus/blur
-  const triggerRef = React.useRef<HTMLButtonElement | HTMLInputElement>(null);
-  const ignoreTriggerBlur = React.useRef(false);
-
-  // popper
-  const popperOptions = {
-    position: 'below' as const,
-    align: 'start' as const,
-    offset: [0, 2] as Offset,
-    ...resolvePositioningShorthand(positioning),
-  };
-
-  const {
-    targetRef: popperTargetRef,
-    containerRef: popperContainerRef,
-  }: {
-    targetRef: React.MutableRefObject<HTMLButtonElement | HTMLInputElement>;
-    containerRef: React.MutableRefObject<HTMLDivElement>;
-  } = usePopper(popperOptions);
-
-  // update value based on selectedOptions
-  const isFirstMount = useFirstMount();
-  const value = React.useMemo(() => {
-    // don't compute value if it is defined through props,
-    if (props.value !== undefined) {
-      return props.value;
-    }
-
-    if (isFirstMount && props.defaultValue !== undefined) {
-      return props.defaultValue;
-    }
-
-    if (multiselect) {
-      return selectedOptions.map(option => option.value).join(', ');
-    }
-
-    return selectedOptions[0]?.value;
-  }, [isFirstMount, multiselect, props.defaultValue, props.value, selectedOptions]);
-
-  const setOpen = (event: ComboboxOpenEvents, newState: boolean) => {
-    onOpenChange?.(event, { open: newState });
-    setOpenState(newState);
-  };
-
-  const onOptionClick = (event: React.MouseEvent<HTMLElement>, option: OptionValue) => {
-    // clicked option should always become active option
-    setActiveOption(getOptionByKey(option.key));
-
-    // close on option click for single-select
-    !multiselect && setOpen(event, false);
-
-    // handle selection change
-    selectOption(event, option);
-  };
-
-  const { primary: triggerNativeProps, root: rootNativeProps } = getPartitionedNativeProps({
-    props,
-    primarySlotTagName: 'button',
-    excludedPropNames: ['children'],
-  });
-
-  const state: ComboboxBaseState = {
-    components: {
-      root: 'div',
-      listbox: Listbox,
-      input: 'button',
-    },
-    root: resolveShorthand(props.root, {
-      required: true,
-      defaultProps: {
-        children: props.children,
-        ...rootNativeProps,
-      },
-    }),
-    listbox: resolveShorthand(props.listbox, {
-      required: true,
-      defaultProps: {
-        ref: popperContainerRef,
-        multiselect,
-        tabIndex: undefined,
-      },
-    }),
-    input: resolveShorthand(props.input, {
-      required: true,
-      defaultProps: {
-        ref: useMergedRefs(ref, triggerRef, popperTargetRef),
-        'aria-expanded': open,
-        'aria-activedescendant': open ? activeOption?.id : undefined,
-        placeholder,
-        value,
-        ...triggerNativeProps,
-      },
-    }),
-    ...optionCollection,
-    activeOption,
-    appearance,
-    inline,
-    onOptionClick,
-    open,
-    options,
-    selectedOptions,
-    size,
-    value,
-  };
-
-  /*
-   * Handle focus when clicking the listbox popup:
-   * 1. Move focus back to the button/input when the listbox is clicked (otherwise it goes to body)
-   * 2. Do not close the listbox on button/input blur when clicking into the listbox
-   */
-  const { onClick: onListboxClick, onMouseDown: onListboxMouseDown } = state.listbox;
-  state.listbox.onClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    triggerRef.current?.focus();
-
-    onListboxClick?.(event);
-  };
-
-  state.listbox.onMouseDown = (event: React.MouseEvent<HTMLDivElement>) => {
-    ignoreTriggerBlur.current = true;
-
-    onListboxMouseDown?.(event);
-  };
-
-  // the trigger should open/close the popup on click or blur
-  const { onBlur: onTriggerBlur, onClick: onTriggerClick, onKeyDown: onTriggerKeyDown } = state.input;
-  state.input.onBlur = (event: React.FocusEvent<HTMLButtonElement> & React.FocusEvent<HTMLInputElement>) => {
-    if (!ignoreTriggerBlur.current) {
-      setOpen(event, false);
-    }
-
-    ignoreTriggerBlur.current = false;
-
-    onTriggerBlur?.(event);
-  };
-
-  state.input.onClick = (event: React.MouseEvent<HTMLButtonElement> & React.MouseEvent<HTMLInputElement>) => {
-    setOpen(event, !open);
-
-    onTriggerClick?.(event);
-  };
-
-  // handle combobox keyboard interaction
-  state.input.onKeyDown = (event: React.KeyboardEvent<HTMLButtonElement> & React.KeyboardEvent<HTMLInputElement>) => {
-    const action = getDropdownActionFromKey(event, { open, multiselect });
-    const maxIndex = count - 1;
-    const activeIndex = activeOption ? getIndexOfKey(activeOption.key) : -1;
-    let newIndex = activeIndex;
-
-    switch (action) {
-      case 'Open':
-        event.preventDefault();
-        setOpen(event, true);
-        break;
-      case 'Close':
-        // stop propagation for escape key to avoid dismissing any parent popups
-        event.stopPropagation();
-        event.preventDefault();
-        setOpen(event, false);
-        break;
-      case 'CloseSelect':
-        !multiselect && setOpen(event, false);
-      // fallthrough
-      case 'Select':
-        activeOption && selectOption(event, activeOption);
-        event.preventDefault();
-        break;
-      case 'Tab':
-        activeOption && selectOption(event, activeOption);
-        break;
-      default:
-        newIndex = getIndexFromAction(action, activeIndex, maxIndex);
-    }
-    if (newIndex !== activeIndex) {
-      // prevent default page scroll/keyboard action if the index changed
-      event.preventDefault();
-      setActiveOption(getOptionAtIndex(newIndex));
-    }
-
-    onTriggerKeyDown?.(event);
-  };
-
-  return state;
-};
