+++ packages/react-components/react-combobox/src/stories/ComboboxDefault.stories.tsx
@@ -10,9 +10,7 @@ export const Default = (props: Partial<ComboboxProps>) => {
       <label id={comboId}>Best pet</label>
       <Combobox aria-labelledby={comboId} placeholder="Select an animal" {...props}>
         {options.map(option => (
-          <Option key={option} disabled={option === 'Ferret'}>
-            {option}
-          </Option>
+          <Option disabled={option === 'Ferret'}>{option}</Option>
         ))}
       </Combobox>
     </>
